Added routes to make transaction form dynamic
