+++ models/ItemVariation.js
@@ -2,21 +2,20 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 const ItemVariationSchema = new Schema({
-    name: {
-        type: String,
-        required: true
-    },
-    gender: {
-        type: [String]
-    },
-    size: {
-        type: [String]
-    },
-    typeColor: {
-        type: [String]
-    }
+  name: {
+    type: String,
+    required: true,
+  },
+  gender: {
+    type: [String],
+  },
+  size: {
+    type: [String],
+  },
+  typeColor: {
+    type: [String],
+  },
 });
 
-const ItemVariation = mongoose.model('ItemVariation', ItemVariationSchema);
-
-module.exports = ItemVariation;
+export default (mongoose.models && mongoose.models.ItemVariation) ? mongoose.models.ItemVariation :
+  mongoose.model('ItemVariation', ItemVariationSchema);
